feat: `main.py` moved out of `src`; `chess` directory in `src`.
