+++ src/chess/__init__.py
@@ -1,0 +1,1 @@
+from chess.chess import Chess
