+++ src/chess/colours.py
@@ -1,7 +1,7 @@
 """Enum classes for the project."""
 # ——————————————————————————————————————————— Imports ——————————————————————————————————————————— #
 # Standard libraries
-from enum import StrEnum, Enum
+from enum import StrEnum
 
 # ———————————————————————————————————————————— Code ———————————————————————————————————————————— #
 
@@ -10,27 +10,3 @@ class Colour(StrEnum):
     """Enum class for piece colours."""
     WHITE = "White"
     BLACK = "Black"
-
-
-class Moves(Enum):
-    """Enum class for moves."""
-    DEFAULT = "default"
-    PIECE_MOVE = "piece move"
-    SHORT_CASTLE = "o-o"
-    LONG_CASTLE = "o-o-o"
-
-
-class Commands(Enum):
-    """Enum class for commands."""
-    YES = "yes"
-    NO = "no"
-    HELP = "help"
-    EXIT = "exit"
-    SAVE_MOVE_HISTORY = "save move history"
-    RESET = "reset"
-    RESIGN = "resign"
-    DRAW = "draw"
-    PRINT_BOARD = "print board"
-
-    # Implement in the future
-    LOAD = "load from move history"
