+++ src/chess/pieces/__init__.py
@@ -1,0 +1,1 @@
+from chess.pieces.pieces import Pawn, Knight, Bishop, Rook, Queen, King
