--- src/pieces/__init__.py
@@ -1,1 +1,0 @@
-from pieces.pieces import Pawn, Knight, Bishop, Rook, Queen, King
