--- src/utils/__init__.py
@@ -1,1 +1,0 @@
-from utils.enums import Colour, Commands, Moves
